requirements.txt
asgiref==3.11.0
Django==6.0.1
django-environ==0.12.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
